Move gestures and refactor code
